src/detection/towers.py
"""Tower detection and status tracking"""

from typing import List
import cv2
import numpy as np
from loguru import logger

from detection.base import BaseDetector
from api.schemas import Tower, Position
from config import settings


class TowerDetector(BaseDetector):
    """
    Detects tower positions and status (alive/destroyed)

    Strategy:
    1. Use known tower positions (fixed locations)
    2. Color detection or template matching
    3. Detect if tower icon is present (alive) or missing (destroyed)
    """

    def __init__(self):
        super().__init__()

        # TODO: Define tower positions for all lanes and tiers

    def initialize(self) -> None:
        """Initialize tower detector"""
        logger.info("Initializing TowerDetector...")
        super().initialize()
        logger.info("TowerDetector ready")

    def detect(self, minimap: np.ndarray) -> List[Tower]:
        """
        Detect all towers and their status

        Args:
            minimap: OpenCV BGR image of minimap

        Returns:
            List of Tower objects
        """
        if minimap is None or minimap.size == 0:
            logger.warning("TowerDetector received empty minimap")
            return []

        towers = []

        try:
            # TODO: Implement tower detection
            # For now, return empty list
            logger.debug("Tower detection not yet implemented")

        except Exception as e:
            logger.error(f"TowerDetector error: {e}")
            return []

        return towers
